Pull-up refactoring of ApplicationBehavior classes
